cvProcessingUsingOpenai/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
boto3==1.35.87
botocore==1.35.87
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
distro==1.9.0
fastapi==0.115.6
fastapi-utils==0.8.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jmespath==1.0.1
lxml==5.3.0
mypy-extensions==1.0.0
openai==1.58.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
psutil==5.9.8
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
